services/service-file-parser/requirements.txt
pika
pydantic
pydantic-settings
pymupdf
python-docx
